requirements.txt
Django==4.1.6
django-environ
psycopg2-binary
discord.py==2.1.1
ipython
lru_dict
parsedatetime
python-dateutil
aiohttp~=3.8.3
enum-properties
django-import-export